requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
streamlit>=1.30.0
gspread>=5.10.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
